Uninstall PIC
